pyproject.toml
[build-system] # Require setuptool version due to https://github.com/pypa/setuptools/issues/2938
requires = ["setuptools>=61.0.0", "wheel"]

 # See https://python-poetry.org/docs/pyproject/ for more keywords for the project table
[project]
 # Name of your package
name = "ffian"
authors = [
    {name = "Marte J. Sætra", email = "martejulie@simula.no"},
    {name = "Ada J. Ellingsrud", email = "ada@simula.no"}
    ]
version = "0.1.0"
description = "Fluid flow in astrocyte networks"
# "readme" the description field in your package, see: https://python-poetry.org/docs/pyproject/#readme
readme = "README.md"
requires-python = ">=3.8"
 # Path to license file, see: https://spdx.org/licenses/ for options
license = {file = "LICENSE"}
# Direct dependencies
dependencies = [
    'numpy',
    'fenics-dolfin'
]

[project.optional-dependencies]
test = [
   "flake8", # Formatting: https://flake8.pycqa.org/en/latest/user/error-codes.html
   "mypy", # Input/Output consistency
   "pytest", # Testing suite
]

docs = [
    "jupyter-book", # Required to build documentation
    "jupytext", # Required to convert .py to .ipynb files
    "ipython!=8.7.0" # https://github.com/ipython/ipython/issues/13845
]

binder = [ # Required to interface with Binder when having a Dockerfile in root
    "jupyterlab",
    "notebook"
]
pypi = [
    "build" # Used for building wheels and uploading to pypi
]

all = ["ffian[test,docs,binder,pypi]"]

[tool.mypy]
ignore_missing_imports = true # Does not show errors when importing untyped libraries
exclude = [ # We only want mypy to consider files that are not generated in installing or building documentation
    "docs/",
    "build/"
]
files = [  # Folder to which files that should be checked by mypy
    "src",
    "test"
]

testpaths = [
    "test"
]


# Pair notebooks with python files
[tool.jupytext]
formats = "demos///ipynb,demos///py:percent"
